requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pygments==2.15.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
